Refactor: Consolidate duplicate conftest.py fixtures
