+++ tests/network_manager/test_darwin.py
@@ -3,6 +3,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
+pytest.importorskip("CoreWLAN")
 pytestmark = pytest.mark.skipif(sys.platform != 'darwin', reason="macOS only")
 
 from vorta.network_status import darwin  # noqa: E402
